Separate extension methods in to specific namespace
Extensions should not polute intellisense namespace. Developers should be
allowed to bring in the extensions they need for the task at hand.

Extensions for different types have been moved in to specific namespaces for
related tasks.
